+++ docs/changes.md
@@ -4,6 +4,7 @@
   * mledit 去掉 scroll_line 属性（感谢俊杰提供补丁)。
   * 修正pages页面设置active属性不生效的问题（感谢俊杰提供补丁)。
   * 修复pages控件的auto_focused属性不生效的问题（感谢俊杰提供补丁)。
+  * 修复list_view_h控件修改item_width属性时没有生效的问题（感谢俊杰提供补丁)。
 
 2025/05/26
   * 修复 tab\_button\_group\_set\_compact 没有立即生效的问题（感谢俊杰提供补丁)。